feat(PlayerService): extract join, disconnect, win-condition logic
refactor(GameService): delegate player lifecycle to PlayerService

test: integrate PlayerService into TestServiceFactory
